refactor: move logic into lib/* files
