+++ index.js
@@ -6,7 +6,10 @@ var Sync = require('pouchdb-hoodie-sync')
 var UnsyncedLocalDocs = require('pouchdb-hoodie-unsynced-local-docs')
 var merge = require('lodash.merge')
 var EventEmitter = require('events').EventEmitter
-var localStorageWrapper = require('humble-localstorage')
+
+var hasLocalChanges = require('./lib/has-local-changes')
+var mapUnsyncedLocalDocs = require('./lib/map-unsynced-local-docs')
+var subscribeToInternalEvents = require('./lib/subscribe-to-internal-events')
 
 PouchDB.plugin({
   hoodieApi: API.hoodieApi,
@@ -27,7 +30,7 @@ function Store (dbName, options) {
     db.hoodieSync({remote: remote, emitter: emitter}),
     db.hoodieApi({emitter: emitter}),
     {
-      findAllUnsynced: mapUnsyncedLocalIds.bind(db),
+      findAllUnsynced: mapUnsyncedLocalDocs.bind(db),
       hasLocalChanges: hasLocalChanges.bind(db)
     }
   )
@@ -47,65 +50,4 @@ Store.defaults = function (defaultOpts) {
   }
 
   return CustomStore
-}
-
-function mapUnsyncedLocalIds (options) {
-  return this.unsyncedLocalDocs.call(this, options)
-
-  .then(function (changedDocs) {
-    return changedDocs.map(function (doc) {
-      doc.id = doc._id
-      delete doc._id
-
-      return doc
-    })
-  })
-}
-
-function hasLocalChanges (objOrId) {
-  var changedIds = localStorageWrapper.getObject('hoodie_changedObjectIds') || []
-  if (objOrId) {
-    var id = objOrId.id ? objOrId.id : objOrId
-    return changedIds.indexOf(id) >= 0
-  }
-
-  return changedIds.length > 0
-}
-
-function markAsChanged (object) {
-  var changedIds = localStorageWrapper.getObject('hoodie_changedObjectIds') || []
-  var id = object.id
-  var hasId = changedIds.indexOf(id) >= 0
-
-  if (hasId) {
-    return
-  }
-  localStorageWrapper.setObject('hoodie_changedObjectIds', changedIds.concat(id))
-}
-
-function unmarkAsChanged (object) {
-  var changedIds = localStorageWrapper.getObject('hoodie_changedObjectIds')
-  var id = object._id
-  var index = changedIds.indexOf(id)
-
-  if (index === -1) {
-    return
-  }
-
-  changedIds.splice(index, 1)
-  localStorageWrapper.setObject('hoodie_changedObjectIds', changedIds)
-}
-
-function subscribeToInternalEvents (emitter) {
-  emitter.on('change', function (eventName, object) {
-    markAsChanged(object)
-  })
-
-  emitter.on('push', function (object) {
-    unmarkAsChanged(object)
-  })
-
-  emitter.on('clear', function () {
-    localStorageWrapper.removeItem('hoodie_changedObjectIds')
-  })
 }
