+++ lib/has-local-changes.js
@@ -1,0 +1,13 @@
+module.exports = hasLocalChanges
+
+var localStorageWrapper = require('humble-localstorage')
+
+function hasLocalChanges (objOrId) {
+  var changedIds = localStorageWrapper.getObject('hoodie_changedObjectIds') || []
+  if (objOrId) {
+    var id = objOrId.id ? objOrId.id : objOrId
+    return changedIds.indexOf(id) >= 0
+  }
+
+  return changedIds.length > 0
+}
